src/pyproject.toml
[tool.poetry]
name = "tessian-platform-interview-challenge"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.3"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
